+++ lib/src/side_effect_provider.dart
@@ -5,6 +5,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 /// An object that provides access to a stream of side effects over time.
 abstract class SideEffectProvider<SideEffect, State>
     extends StateStreamable<State> {
-  /// The current [sideEffects] of side effects.
+  /// The stream of side effects.
   Stream<SideEffect> get sideEffects;
 }